@@ -4,13 +4,14 @@ description: Just Another Workout Timer
 version: 1.6.0+20210311
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.13.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_tts:
   path_provider:
+  flutter_file_dialog:
   pref:
   fluttertoast:
   prefs:
