@@ -1,5 +1,7 @@
 import 'dart:io';
+import 'dart:typed_data';
 
+import 'package:flutter_file_dialog/flutter_file_dialog.dart';
 import 'package:path_provider/path_provider.dart';
 
 import 'utils.dart';
@@ -7,7 +9,6 @@ import 'workout.dart';
 
 Future<String> get _localPath async {
   final directory = await getExternalStorageDirectory();
-
   await Directory('${directory!.path}/workouts').create();
 
   return directory.path;
@@ -18,9 +19,38 @@ Future<File> _loadWorkoutFile(String title) async {
   return File('$path/workouts/${Utils.removeSpecialChar(title)}.json');
 }
 
+Future<void> exportWorkout(String title) async {
+  var workout = await loadWorkout(title: title);
+  var backup = Backup(workouts: [workout]);
+  final params = SaveFileDialogParams(
+      data: Uint8List.fromList(backup.toRawJson().codeUnits),
+      fileName: '${Utils.removeSpecialChar(title)}.json');
+  await FlutterFileDialog.saveFile(params: params);
+}
+
+Future<void> exportAllWorkouts() async {
+  var backup = Backup(workouts: await getAllWorkouts());
+  final params = SaveFileDialogParams(
+      data: Uint8List.fromList(backup.toRawJson().codeUnits),
+      fileName: 'Backup.json');
+  await FlutterFileDialog.saveFile(params: params);
+}
+
+Future<void> importBackup() async {
+  final params = OpenFileDialogParams(
+      dialogType: OpenFileDialogType.document,
+      fileExtensionsFilter: ['json'],
+      allowEditing: false);
+  final filePath = await FlutterFileDialog.pickFile(params: params);
+  if (filePath != null && filePath.isNotEmpty) {
+    var backup = await File(filePath).readAsString();
+    var workouts = Backup.fromRawJson(backup).workouts;
+    workouts.forEach(writeWorkout);
+  }
+}
+
 void writeWorkout(Workout workout) async {
   final file = await _loadWorkoutFile(workout.title);
-
   file.writeAsString(workout.toRawJson(), flush: true);
 }
 
@@ -29,8 +59,8 @@ Future<bool> workoutExists(String title) async {
   return file.exists();
 }
 
-Future<Workout> loadWorkout(String title) async {
-  final file = await _loadWorkoutFile(title);
+Future<Workout> loadWorkout({String? title, File? workoutFile}) async {
+  final file = workoutFile ?? await _loadWorkoutFile(title!);
   var contents = await file.readAsString();
 
   return Workout.fromRawJson(contents);
@@ -62,5 +92,6 @@ Future<List<Workout>> getAllWorkouts() async {
       .listSync()
       .map((e) => e.path.split("/").last.split(".").first)
       .toList();
-  return await Future.wait(titles.map((t) async => await loadWorkout(t)));
+  return await Future.wait(
+      titles.map((t) async => await loadWorkout(title: t)));
 }
